Output non-fizzbuzz numbers
This creates a Fizzbuzz class that implements an output function. This
function will now correctly implement the first two integers, which are not
'fizz' or 'buzz'. i.e. '1, 2'

Implemented without using an enumerable or maps, etc. I want to be able
to run with an argument of 10000 if I want. Therefore, I don't want the
whole thing in memory.